@@ -10,21 +10,20 @@ use CodeIgniter\Config\BaseConfig;
 class Feature extends BaseConfig
 {
     /**
-     * Enable multiple filters for a route or not.
-     *
-     * If you enable this:
-     *   - CodeIgniter\CodeIgniter::handleRequest() uses:
-     *     - CodeIgniter\Filters\Filters::enableFilters(), instead of enableFilter()
-     *   - CodeIgniter\CodeIgniter::tryToRouteIt() uses:
-     *     - CodeIgniter\Router\Router::getFilters(), instead of getFilter()
-     *   - CodeIgniter\Router\Router::handle() uses:
-     *     - property $filtersInfo, instead of $filterInfo
-     *     - CodeIgniter\Router\RouteCollection::getFiltersForRoute(), instead of getFilterForRoute()
-     */
-    public bool $multipleFilters = false;
-
-    /**
      * Use improved new auto routing instead of the default legacy version.
      */
     public bool $autoRoutesImproved = true;
+
+    /**
+     * Use filter execution order in 4.4 or before.
+     */
+    public bool $oldFilterOrder = false;
+
+    /**
+     * The behavior of `limit(0)` in Query Builder.
+     *
+     * If true, `limit(0)` returns all records. (the behavior of 4.4.x or before in version 4.x.)
+     * If false, `limit(0)` returns no records. (the behavior of 3.1.9 or later in version 3.x.)
+     */
+    public bool $limitZeroAsAll = true;
 }
